Support latest service-catalog in OS 3.7
Update the ansible service broker setup to wait for the service-catalog
to be available. This prevents a race condition when we are working with
coalmine and the service-catalog (giving it time to redeploy).

Also, with the service-catalog being pushed into `kube-service-catalog`
update the openshift_setup to register the project where the service
catalog was put.